+++ src/phylum/install/__init__.py
@@ -1,0 +1,4 @@
+"""Package for the phylum install script."""
+
+# Dynamically create the script name based on the package structure to help stay DRY
+SCRIPT_NAME = __name__.replace(".", "-")
