+++ src/phylum/install/__main__.py
@@ -1,0 +1,3 @@
+from phylum.install.cli import main
+
+main()
