+++ tests/functional/test_install.py
@@ -1,11 +1,12 @@
-""""Test the command line interface (CLI)."""
+""""Test the phylum-install command line interface (CLI)."""
 
 import subprocess
 import sys
 
-from phylum import PKG_NAME, __version__
+from phylum import __version__
+from phylum.install import SCRIPT_NAME
 
-from .constants import PYPROJECT
+from ..constants import PYPROJECT
 
 
 def test_run_as_module():
@@ -15,7 +16,7 @@ def test_run_as_module():
     NOTE: The <module_name> must be specified with an underscore, even if the corresponding
           script entry point is specified with a dash.
     """
-    cmd_line = [sys.executable, "-m", "phylum", "--help"]
+    cmd_line = [sys.executable, "-m", "phylum.install", "--help"]
     ret = subprocess.run(cmd_line)
     assert ret.returncode == 0, "Running the package as a module failed"
 
@@ -24,15 +25,15 @@ def test_run_as_script():
     """Ensure the CLI can be called by it's script entry point."""
     scripts = PYPROJECT.get("tool", {}).get("poetry", {}).get("scripts", {})
     assert scripts, "There should be at least one script entry point"
-    for script in scripts:
-        ret = subprocess.run([script, "-h"])
-        assert ret.returncode == 0, f"{script} entry point failed"
+    assert SCRIPT_NAME in scripts, "The phylum-install script should be a defined entry point"
+    ret = subprocess.run([SCRIPT_NAME, "-h"])
+    assert ret.returncode == 0, f"{SCRIPT_NAME} entry point failed"
 
 
 def test_version_option():
     """Ensure the correct program name and version is displayed when using the `--version` option."""
     # The argparse module adds a newline to the output
-    expected_output = f"{PKG_NAME} {__version__}\n"
+    expected_output = f"{SCRIPT_NAME} {__version__}\n"
     cmd_line = [sys.executable, "-m", "phylum", "--version"]
     ret = subprocess.run(cmd_line, check=True, capture_output=True, encoding="utf-8")
     assert ret.stdout == expected_output, "Output did not match expected input"
